chore: move depl_calls_svc plugin
